Moved test files out of test folder because there were issues with the build when trying to run tests there
